refactor(frontend): move logout button and hook from app/_components to components/nav
