+++ frontend/src/components/nav/app-shell.test.tsx
@@ -28,7 +28,7 @@ vi.mock("next/navigation", () => ({
 }));
 
 // LogoutButton reaches into Supabase. Stub it to keep the test self-contained.
-vi.mock("@/app/_components/logout-button", () => ({
+vi.mock("./logout-button", () => ({
   LogoutButton: () => (
     <button type="button" data-testid="logout-button">
       Sign out
